RavenDB-17068 Moved mismatches handling to separate file
