+++ src/Raven.Server/Documents/Indexes/Static/CurrentIndexingScope.cs
@@ -36,21 +36,16 @@ namespace Raven.Server.Documents.Indexes.Static
         /// [collection: [key: [referenceKeys]]]
         public Dictionary<string, Dictionary<Slice, HashSet<Slice>>> ReferencesByCollectionForCompareExchange;
 
+        /*
         public Dictionary<string, Dictionary<string, LoadFailure>> MismatchedReferences;
 
         private const int MaxMismatchedReferencesPerSource = 10;
         private const int MaxMismatchedDocumentLoadsPerIndex = 10;
         
         private bool _lastLoadMismatched = false;
+        */
+        internal MismatchedReferencesHandler _mismatchedReferencesHandler = new MismatchedReferencesHandler();
 
-        public class LoadFailure
-        {
-            public string SourceId;
-            public string ReferenceId;
-            public string ActualCollection;
-            public HashSet<string> MismatchedCollections;
-        }
-        
         [ThreadStatic]
         public static CurrentIndexingScope Current;
 
@@ -238,82 +233,25 @@ namespace Raven.Server.Documents.Indexes.Static
                 {
                     if (string.Equals(collection, collectionName, StringComparison.OrdinalIgnoreCase) == false)
                     {
-                        MismatchedReferences ??= new Dictionary<string, Dictionary<string, LoadFailure>>();
-
-                        if (MismatchedReferences.Count < MaxMismatchedDocumentLoadsPerIndex)
+                        if (_mismatchedReferencesHandler.MismatchedReferences.Count < MismatchedReferencesHandler.MaxMismatchedDocumentLoadsPerIndex)
                         {
-                            _lastLoadMismatched = true;
-                            HandleMismatchedReference(document, collectionName, id, collection);
+                            _mismatchedReferencesHandler._lastLoadMismatched = true;
+                            _mismatchedReferencesHandler.HandleMismatchedReference(document, collectionName, id, collection);
                         }
 
                         return DynamicNullObject.Null;
                     }
 
-                    if (_lastLoadMismatched)
+                    if (_mismatchedReferencesHandler._lastLoadMismatched)
                     {
-                        RemoveMismatchedReferenceOnMatchingLoad(document, id);
-                        _lastLoadMismatched = false;
+                        _mismatchedReferencesHandler.RemoveMismatchedReferenceOnMatchingLoad(document, id);
+                        _mismatchedReferencesHandler._lastLoadMismatched = false;
                     }
                 }
 
                 // we can't share one DynamicBlittableJson instance among all documents because we can have multiple LoadDocuments in a single scope
                 return new DynamicBlittableJson(document);
             }
-        }
-
-        private void HandleMismatchedReference(Document referencedDocument, string referencedCollectionName, LazyStringValue sourceId, string actualCollection)
-        {
-            // another mismatch for source document
-            if(MismatchedReferences.TryGetValue(sourceId, out Dictionary<string, LoadFailure> mismatchesForDocument) && mismatchesForDocument.Count < MaxMismatchedReferencesPerSource)
-            {
-                // another mismatch referencing the same document
-                if (mismatchesForDocument.TryGetValue(referencedDocument.Id, out LoadFailure loadFailure))
-                    loadFailure.MismatchedCollections.Add(referencedCollectionName);
-                else
-                {
-                    mismatchesForDocument.Add(
-                        referencedDocument.Id, new LoadFailure()
-                        {
-                            SourceId = sourceId, 
-                            ReferenceId = referencedDocument.Id,
-                            ActualCollection = actualCollection,
-                            MismatchedCollections = new HashSet<string>()
-                            {
-                                referencedCollectionName
-                            } 
-                        });
-                }
-            }
-            else
-            {
-                // first mismatch for source document
-                LoadFailure failure = new ()
-                {
-                    SourceId = sourceId, 
-                    ReferenceId = referencedDocument.Id,
-                    ActualCollection = actualCollection,
-                    MismatchedCollections = new HashSet<string>()
-                    {
-                        referencedCollectionName
-                    }
-                };
-                            
-                MismatchedReferences.Add(sourceId, new Dictionary<string, LoadFailure>(){ {referencedDocument.Id, failure} });
-            }
-        }
-
-        private void RemoveMismatchedReferenceOnMatchingLoad(Document document, string sourceId)
-        {
-            if (MismatchedReferences.TryGetValue(sourceId, out var failing) == false) 
-                return;
-            
-            failing.Remove(document.Id);
-
-            if (failing.Count == 0)
-                MismatchedReferences.Remove(sourceId);
-
-            if (MismatchedReferences.Count == 0)
-                MismatchedReferences = null;
         }
 
         public unsafe dynamic LoadCompareExchangeValue(LazyStringValue keyLazy, string keyString)
